requirements.txt
Flask
moviepy==1.0.3
pydub
openai-whisper
python-dotenv